app-react: colocate nav boundary test
